apps/api/pyproject.toml
[project]
name = "amilyhub-exporter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "xlrd>=2.0.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
